Prevent kubectl override of exec plugin credentials.
This is done by never storing the retrieved credentials in
`KUBECONFIG` in the first place. Another bonus of using this approach
is that it's possible to wipe the `~/.kube/config` file, restore it and
still have the previously retrieved credential to be used.

Quick fix for
[issue #4](https://github.com/Bisnode/kubectl-login/issues/4)
which in turn is a workaround for
[this issue](https://github.com/kubernetes/kubernetes/issues/87369)